Rename IRCF and remove module Position
I was confused about [ICRF]. First, the name was wrong: ICRF and not
IRCF.

But most importantly, while `Geometric` and `Astrometric` are ICRF
positions, it will not necessarily be the case of future positions (`MeanOfDate`
and `Apparent`, coming soon).

This renames the previous class into `Position` so that all inherit from
it without having to be ICRF positions.

It also removes the module `Position` for clarity.

[ICRF]: https://en.wikipedia.org/wiki/International_Celestial_Reference_System_and_its_realizations
