+++ lib/astronoby.rb
@@ -40,9 +40,9 @@ require "astronoby/moon_phase"
 require "astronoby/nutation"
 require "astronoby/observer"
 require "astronoby/precession"
-require "astronoby/position/ircf"
-require "astronoby/position/geometric"
-require "astronoby/position/astrometric"
+require "astronoby/position"
+require "astronoby/positions/geometric"
+require "astronoby/positions/astrometric"
 require "astronoby/refraction"
 require "astronoby/time/greenwich_sidereal_time"
 require "astronoby/time/local_sidereal_time"
